chore(add-asset): prop type validation & css
- add prop type validation to components
- add css styling
