+++ src/_css/AddAssetComponent.css
@@ -1,0 +1,8 @@
+.form-wrapper {
+    width: 70%;
+}
+
+.label-style {
+    width: 100px !important;
+    text-align: left !important;
+}
