+++ src/components/AddAsset/AddAssetDropdown.jsx
@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import * as React from 'react';
 import { Dropdown } from 'semantic-ui-react';
 
@@ -14,4 +15,12 @@ const AddAssetDropdown = (props) => (
     onChange={props.onChange}
   />
 );
+
+AddAssetDropdown.propTypes = {
+  label: PropTypes.string,
+  options: PropTypes.array,
+  placeholder: PropTypes.string,
+  onChange: PropTypes.func
+};
+
 export default AddAssetDropdown;
